agapatch: fix VC compilation errors

+++ Library/Implementation/HTFile.c
@@ -422,7 +422,7 @@ HTFormat HTFileFormat(const char* filename, HTAtom** pencoding) {
 	HTSuffix* suff;
 	int n;
 	int i;
-	int lf = strlen(filename);
+	int lf = (int) strlen(filename);
 
 #ifndef NO_INIT
 	if(!HTSuffixes) HTFileInit();
@@ -432,7 +432,7 @@ HTFormat HTFileFormat(const char* filename, HTAtom** pencoding) {
 	for(i = 0; i < n; i++) {
 		int ls;
 		suff = HTList_objectAt(HTSuffixes, i);
-		ls = strlen(suff->suffix);
+		ls = (int) strlen(suff->suffix);
 		if((ls <= lf) && 0 == strcmp(suff->suffix, filename + lf - ls)) {
 			int j;
 			*pencoding = suff->encoding;
@@ -441,7 +441,7 @@ HTFormat HTFileFormat(const char* filename, HTAtom** pencoding) {
 			for(j = 0; j < n; j++) {  /* Got encoding, need representation */
 				int ls2;
 				suff = HTList_objectAt(HTSuffixes, j);
-				ls2 = strlen(suff->suffix);
+				ls2 = (int) strlen(suff->suffix);
 				if((ls <= lf) && 0 == strncmp(
 						suff->suffix, filename + lf - ls - ls2, ls2)) {
 					if(suff->rep) return suff->rep;
@@ -474,7 +474,7 @@ float HTFileValue(const char* filename) {
 	HTSuffix* suff;
 	int n;
 	int i;
-	int lf = strlen(filename);
+	int lf = (int) strlen(filename);
 
 #ifndef NO_INIT
 	if(!HTSuffixes) HTFileInit();
@@ -483,7 +483,7 @@ float HTFileValue(const char* filename) {
 	for(i = 0; i < n; i++) {
 		int ls;
 		suff = HTList_objectAt(HTSuffixes, i);
-		ls = strlen(suff->suffix);
+		ls = (int) strlen(suff->suffix);
 		if((ls <= lf) && 0 == strcmp(suff->suffix, filename + lf - ls)) {
 			if(TRACE) {
 				fprintf(
