+++ Library/Implementation/HTWSRC.c
@@ -294,7 +294,7 @@ static void WSRC_gen_html(HTStream* me, HTBool source_file) {
 		char* shortname = 0;
 		int l;
 		StrAllocCopy(shortname, me->par_value[PAR_DATABASE_NAME]);
-		l = strlen(shortname);
+		l = (int) strlen(shortname);
 		if(l > 4 && !strcasecomp(shortname + l - 4, ".src")) {
 			shortname[l - 4] = 0;    /* Chop of .src -- boring! */
 		}
